Plugin improvements
* Removed second temporary plugin
* Moved logic from `sample`'s build.gradle to the plugin
