--- buildSrc/src/main/kotlin/MockingbirdTestPlugin.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright Careem, an Uber Technologies Inc. company
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-
-// TODO test to add sourceSET ONLY
-abstract class MockingbirdTestPlugin : Plugin<Project> {
-
-    override fun apply(target: Project) {
-        try {
-            // TODO delete file before build
-            configureSourceSets(target)
-            println("generated path added to source sets")
-        } catch (e: Exception) {
-            // Useful to debug
-            e.printStackTrace()
-            throw e
-        }
-    }
-
-    private fun configureSourceSets(target: Project) {
-        // TODO check if kmpProject before this
-        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            sourceSets.getByName("commonTest") {
-                kotlin.srcDir("build/generated/mockingbird")
-            }
-        }
-    }
-}
