+++ samples/build.gradle
@@ -60,15 +60,5 @@ kotlin {
     }
 }
 
-
-// Part below this needs to be extracted and made generic
-apply plugin: MockingbirdTestPlugin
-task generateMocks() {
-    dependsOn tasks.assemble
-    doLast{
-        apply plugin: MockingbirdPlugin
-    }
-}
-
-allTests.dependsOn generateMocks
+apply plugin: MockingbirdPlugin
